.history/requirements_20260421134722.txt
# Core data science
pandas
numpy
scipy
scikit-learn
joblib

# HTTP / scraping
requests
beautifulsoup4

# MLB data (free)
pybaseball
mlbstatsapi

# Visualization
matplotlib
seaborn

# Utilities
python-dotenv
tabulate
jupyter